fix anndata bug
anndata v 0.12.0 changed behavior so that an AnnData with X=None
still writes the dataset 'X' to an h5ad file when you call
write_h5ad. This broke some of our utilities for slicing and
dicing h5ad files. This commit fixes the bug by deleting 'X'
whenever it exists but we do not expect it to. This will probably
introduce some overhead in the h5ad files, but it is minimal, since
it only occurs in cases where X=None anyway.